Update model, add migration for vn tables

+++ API/SeiyuuMoe.Domain/Entities/AnimeRoleType.cs
@@ -17,7 +17,7 @@ namespace SeiyuuMoe.Domain.Entities
 		public virtual ICollection<AnimeRole> Role { get; set; }
 	}
 
-	public enum AnimeRoleTypeId: int
+	public enum AnimeRoleTypeId: long
 	{
 		All = 0,
 		Main = 1,
